# openapi.yaml
openapi: 3.1.0
info:
  title: PrimeSoft Alliance Unified API
  version: 1.0.0
  description: >
    Production-grade, API-first SaaS platform for AI chatbots, bookings, contacts, forms,
    content management, multimedia, and analytics.
servers:
  - url: "https://{domain}/v1"
    description: "Production Environment (Render or custom domain)"
    variables:
      domain:
        default: "your-app.onrender.com"
        description: "The host domain of your application"
  - url: "http://localhost:3000/v1"
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /bookings:
    get:
      summary: List Bookings
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
        - $ref: '#/components/parameters/QueryPage'
      responses:
        '200':
          description: Paginated bookings
    post:
      summary: Create a Booking
      tags: [Bookings]
      parameters:
        - $ref: '#/components/parameters/HeaderIdempotencyKey'
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '201':
          description: Booking created
  /bookings/{booking_id}:
    get:
      summary: Get Booking Details
      tags: [Bookings]
      parameters:
        - name: booking_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Booking details
    patch:
      summary: Update Booking
      tags: [Bookings]
      parameters:
        - name: booking_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete Booking
      tags: [Bookings]
      parameters:
        - name: booking_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Deleted
...
  /forms:
    get:
      summary: List all forms
      tags: [Forms]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Paginated forms
  /forms/{form_id}:
    get:
      summary: Get form details
      tags: [Forms]
      parameters:
        - name: form_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Form details
    patch:
      summary: Update form
      tags: [Forms]
      parameters:
        - name: form_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Form updated
    delete:
      summary: Delete form
      tags: [Forms]
      parameters:
        - name: form_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Form deleted
...
  /analytics/bookings:
    get:
      summary: Bookings analytics
      tags: [Analytics]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Bookings Analytics
...
  /analytics/forms:
    get:
      summary: Forms analytics
      tags: [Analytics]
      parameters:
        - $ref: '#/components/parameters/HeaderClientId'
      responses:
        '200':
          description: Forms Analytics
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    HeaderClientId:
      name: x-client-id
      in: header
      required: true
      schema: { type: string }
    HeaderIdempotencyKey:
      name: idempotency-key
      in: header
      required: false
      schema: { type: string }
    QueryPage:
      name: page
      in: query
      schema: { type: integer, default: 1 }
  schemas:
...
    Booking:
      type: object
      properties:
        serviceId: { type: string }
        customerName: { type: string }
        email: { type: string, format: email }
        date: { type: string, format: date }
        time: { type: string }
...